various refactors
