+++ src/interfaces/CheckerCalendar.ts
@@ -1,0 +1,4 @@
+export interface CheckerCalendarItem {
+  day: string;
+  value: number;
+}
