+++ src/interfaces/LineData.ts
@@ -1,0 +1,10 @@
+export interface CheckerCords {
+  x: string;
+  y: number;
+}
+
+export interface LineData {
+  id: string;
+  color: string;
+  data: CheckerCords[];
+}
