+++ src/settings/chart/chartColors.ts
@@ -1,0 +1,1 @@
+export const lineColor = 'hsl(192, 70%, 50%)';
