+++ src/settings/chart/commonProperties.ts
