+++ src/utils/mappings/checkerCalendar.ts
@@ -1,0 +1,8 @@
+import { Checker, CheckerFields } from '../../interfaces/Checker';
+import { CheckerCalendarItem } from '../../interfaces/CheckerCalendar';
+
+export const transformCheckersToCalendarFormat = (checkers: Checker[]): CheckerCalendarItem[] => {
+  return checkers.map(checker => {
+    return { day: checker[CheckerFields.created], value: checker[CheckerFields.amount] };
+  });
+};
